utils/requirements.txt
ansible==2.9.6
apache-libcloud==2.8.0
appdirs==1.4.4
argcomplete==1.12.1
astroid==2.4.2
attrs==19.3.0
Automat==0.8.0
binary==1.0.0
black==20.8b1
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
entrypoints==0.3
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==16.10.1
isort==5.7.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.23.0
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.1
pexpect==4.6.0
pipx==0.15.5.1
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pykerberos==1.1.14
pylint==2.6.0
pymacaroons==0.13.0
pymdown-extensions==9.5
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypinfo==21.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-frontmatter==1.0.0
pywinrm==0.3.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.28.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
selinux==3.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
tinydb==4.7.0
tinyrecord==0.2.0
toml==0.10.2
Twisted==18.9.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.11
userpath==1.4.1
wadllib==1.3.3
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.8.1
zope.interface==4.7.1
